task/multiple-crc-files-in-lib-jaia-1968 (#1158)
* Move crc header files into the src/lib/crc directory, and move the serial header files into src/lib/serial

* Remove unused crc16 code.  Rename mis-named SerialThreadCRC32 class to SerialThreadFletcher16 for the Arduino comms.  Move Fletcher-16 code to fletcher16.h

* Add note to implement fletcher-16 on incoming data in the SerialThreadFletcher16 class.

* Remove unused boost crc

---------
