--- src/arduino/jaiabot_runtime/crc16.h
@@ -1,1 +1,0 @@
-../../lib/lora/crc16.h
