+++ src/bin/sensors/app.cpp
@@ -20,8 +20,6 @@
 // You should have received a copy of the GNU General Public License
 // along with the Jaia Binaries.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <boost/crc.hpp>
-
 #include <goby/middleware/marshalling/protobuf.h>
 // this space intentionally left blank
 #include <goby/middleware/io/cobs/serial.h>
@@ -36,7 +34,6 @@
 #include "jaiabot/crc/crc32.h"
 #include "jaiabot/groups.h"
 #include "jaiabot/messages/health.pb.h"
-#include "jaiabot/messages/sensor/catalog.pb.h"
 #include "jaiabot/messages/sensor/sensor_core.pb.h"
 
 using goby::glog;
@@ -74,7 +71,6 @@ class Sensors : public zeromq::MultiThreadApplication<config::Sensors>
         initialization_error_names;
     std::map<jaiabot::sensor::protobuf::Sensor, jaiabot::protobuf::Warning>
         initialization_warning_names;
-    boost::crc_32_type crc32_calc_;
 };
 
 } // namespace apps
