--- src/lib/messages/lora_test.proto
@@ -1,53 +1,0 @@
-syntax = "proto2";
-
-import "dccl/option_extensions.proto";
-import "jaiabot/messages/feather.proto";
-import "goby/middleware/protobuf/gpsd.proto";
-import "jaiabot/messages/geographic_coordinate.proto";
-
-package jaiabot.protobuf;
-
-message LoRaTestData
-{
-    option (dccl.msg) = {
-        id: 124
-        max_bytes: 64
-        codec_version: 3
-        unit_system: "si"
-    };
-
-    required uint64 time = 1 [(dccl.field) = {
-        codec: "dccl.time2"
-        units { prefix: "micro" derived_dimensions: "time" }
-    }];
-
-    required uint32 index = 2 [(dccl.field) = { min: 0 max: 1000 }];
-
-    required GeographicCoordinate location = 10;
-}
-
-message LoRaReport
-{
-    option (dccl.msg) = {
-        unit_system: "si"
-    };
-
-    enum Status
-    {
-        GOOD_RECEPTION = 1;
-        NO_PACKET = 2;
-    }
-
-    required uint64 time = 1 [(dccl.field) = {
-        codec: "dccl.time2"
-        units { prefix: "micro" derived_dimensions: "time" }
-    }];
-
-    required Status status = 10;
-    optional LoRaMessage feather_msg = 20;
-    optional LoRaTestData test_data = 30;
-    optional goby.middleware.protobuf.gpsd.TimePositionVelocity gps_tpv = 40;
-    optional double range = 50
-        [(dccl.field) = { units { base_dimensions: "L" } }];
-    optional double packet_success = 60;
-}
