+++ src/lib/serial/serial_fletcher16.h
@@ -9,7 +9,7 @@
 #include "goby/middleware/io/detail/serial_interface.h" // for SerialThread
 #include "nanopb/jaiabot/messages/feather.pb.h"
 
-#include "crc16.h"
+#include "jaiabot/crc/fletcher16.h"
 
 namespace goby
 {
@@ -52,7 +52,7 @@ template <const goby::middleware::Group& line_in_group,
               goby::middleware::io::PubSubLayer::INTERTHREAD,
           template <class> class ThreadType = goby::middleware::SimpleThread,
           bool use_indexed_groups = false>
-class SerialThreadCRC32
+class SerialThreadFletcher16
     : public goby::middleware::io::detail::SerialThread<line_in_group, line_out_group,
                                                         publish_layer, subscribe_layer, ThreadType,
                                                         use_indexed_groups>
@@ -62,12 +62,12 @@ class SerialThreadCRC32
                                                    subscribe_layer, ThreadType, use_indexed_groups>;
 
   public:
-    SerialThreadCRC32(const goby::middleware::protobuf::SerialConfig& config, int index = -1)
+    SerialThreadFletcher16(const goby::middleware::protobuf::SerialConfig& config, int index = -1)
         : Base(config, index)
     {
     }
 
-    ~SerialThreadCRC32() {}
+    ~SerialThreadFletcher16() {}
 
   private:
     void async_read() override;
@@ -135,8 +135,9 @@ template <const goby::middleware::Group& line_in_group,
           goby::middleware::io::PubSubLayer publish_layer,
           goby::middleware::io::PubSubLayer subscribe_layer, template <class> class ThreadType,
           bool use_indexed_groups>
-void jaiabot::lora::SerialThreadCRC32<line_in_group, line_out_group, publish_layer, subscribe_layer,
-                                      ThreadType, use_indexed_groups>::async_read()
+void jaiabot::lora::SerialThreadFletcher16<line_in_group, line_out_group, publish_layer,
+                                           subscribe_layer, ThreadType,
+                                           use_indexed_groups>::async_read()
 {
     buffer_write_ptr_ = buffer_.data();
     read_first_byte();
@@ -147,8 +148,9 @@ template <const goby::middleware::Group& line_in_group,
           goby::middleware::io::PubSubLayer publish_layer,
           goby::middleware::io::PubSubLayer subscribe_layer, template <class> class ThreadType,
           bool use_indexed_groups>
-void jaiabot::lora::SerialThreadCRC32<line_in_group, line_out_group, publish_layer, subscribe_layer,
-                                      ThreadType, use_indexed_groups>::read_first_byte()
+void jaiabot::lora::SerialThreadFletcher16<line_in_group, line_out_group, publish_layer,
+                                           subscribe_layer, ThreadType,
+                                           use_indexed_groups>::read_first_byte()
 {
     boost::asio::async_read(
         this->mutable_serial_port(),
@@ -183,8 +185,9 @@ template <const goby::middleware::Group& line_in_group,
           goby::middleware::io::PubSubLayer publish_layer,
           goby::middleware::io::PubSubLayer subscribe_layer, template <class> class ThreadType,
           bool use_indexed_groups>
-void jaiabot::lora::SerialThreadCRC32<line_in_group, line_out_group, publish_layer, subscribe_layer,
-                                      ThreadType, use_indexed_groups>::read_magic()
+void jaiabot::lora::SerialThreadFletcher16<line_in_group, line_out_group, publish_layer,
+                                           subscribe_layer, ThreadType,
+                                           use_indexed_groups>::read_magic()
 {
     boost::asio::async_read(
         this->mutable_serial_port(),
@@ -221,8 +224,9 @@ template <const goby::middleware::Group& line_in_group,
           goby::middleware::io::PubSubLayer publish_layer,
           goby::middleware::io::PubSubLayer subscribe_layer, template <class> class ThreadType,
           bool use_indexed_groups>
-void jaiabot::lora::SerialThreadCRC32<line_in_group, line_out_group, publish_layer, subscribe_layer,
-                                      ThreadType, use_indexed_groups>::read_size()
+void jaiabot::lora::SerialThreadFletcher16<line_in_group, line_out_group, publish_layer,
+                                           subscribe_layer, ThreadType,
+                                           use_indexed_groups>::read_size()
 {
     boost::asio::async_read(
         this->mutable_serial_port(),
@@ -264,8 +268,9 @@ template <const goby::middleware::Group& line_in_group,
           goby::middleware::io::PubSubLayer publish_layer,
           goby::middleware::io::PubSubLayer subscribe_layer, template <class> class ThreadType,
           bool use_indexed_groups>
-void jaiabot::lora::SerialThreadCRC32<line_in_group, line_out_group, publish_layer, subscribe_layer,
-                                      ThreadType, use_indexed_groups>::read_body()
+void jaiabot::lora::SerialThreadFletcher16<line_in_group, line_out_group, publish_layer,
+                                           subscribe_layer, ThreadType,
+                                           use_indexed_groups>::read_body()
 {
     boost::asio::async_read(
         this->mutable_serial_port(),
@@ -275,6 +280,8 @@ void jaiabot::lora::SerialThreadCRC32<line_in_group, line_out_group, publish_lay
         [this](const boost::system::error_code& ec, std::size_t bytes_transferred) {
             if (!ec && bytes_transferred > 0)
             {
+                // Need to check fletcher-16 checksum for message integrity here
+
                 buffer_write_ptr_ += bytes_transferred;
                 auto io_msg = std::make_shared<goby::middleware::protobuf::IOData>();
                 io_msg->set_data(std::string(buffer_.data(), buffer_write_ptr_));
